+++ simple_tree_fractal/simple_tree.h
@@ -5,4 +5,4 @@
 
 std::vector<Line> iterate(const std::vector<Line>& begin_lines, std::vector<Line> &all_lines, float scaling);
 
-std::vector<Line> simulate(int iterations, float scaling, float window_height);
+std::vector<Line> simulate(int iterations, float scaling, int window_height);
